+++ CMakeLists.txt
@@ -50,12 +50,17 @@ add_library(bra_fs_c STATIC src/bra_fs_c.cpp)
 target_link_libraries(bra_fs_c PRIVATE bra_fs)
 
 add_library(lib_bra STATIC src/lib_bra.c)
+target_sources(lib_bra
+    PUBLIC
+        src/lib_bra_io.c
+        src/encoders/bra_rle.c
+)
 target_include_directories(lib_bra
     PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_BINARY_DIR}/src> # for configure_file (version.h)
         ${CMAKE_CURRENT_SOURCE_DIR}/src
 )
-target_link_libraries(lib_bra PRIVATE bra_log bra_fs_c)
+target_link_libraries(lib_bra PUBLIC bra_log bra_fs_c)
 set_target_properties(lib_bra PROPERTIES
     PREFIX ""
 )
