+++ src/lib_bra_types.h
@@ -15,3 +15,9 @@ typedef enum bra_fs_overwrite_policy_e
     BRA_OVERWRITE_ALWAYS_YES = 1,
     BRA_OVERWRITE_ALWAYS_NO  = 2,
 } bra_fs_overwrite_policy_e;
+
+typedef struct bra_rle_chunk_t
+{
+    bra_rle_counts_t counts;    //!< counts is stored as -1, i.e. 0 means 1 and 255 means 256
+    char             value;     //!< the repeated char.
+} bra_rle_chunk_t;
