+++ test/test_bra_encoders.cpp
@@ -1,5 +1,6 @@
 #include "bra_test.hpp"
 
+// #include <encoders/bra_rle.h>
 #include <lib_bra.h>
 
 ///////////////////////////////////////////////////////////////////////////////
